pyproject.toml
[project]
name = "streamlit-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "koreanize-matplotlib>=0.1.1",
    "openpyxl>=3.1.5",
    "seaborn>=0.13.2",
    "streamlit>=1.55.0",
    "streamlit-lottie>=0.0.5",
]
